requirements_jairsan_atSamaro.txt
atari-py==0.1.1
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
gym==0.9.2
html5lib==0.9999999
idna==2.5
Markdown==2.6.8
matplotlib==2.0.2
numpy==1.13.1
olefile==0.44
opencv-python==3.2.0.8
Pillow==4.2.1
protobuf==3.3.0
pyglet==1.2.4
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.2
scipy==0.19.1
six==1.10.0
tensorflow-gpu==1.2.1
tflearn==0.3.2
urllib3==1.22
Werkzeug==0.12.2
